Extract partial for delegated delivery code
